refactor(gateway): extract isFreeModel from models.ts
models.ts is reachable from client bundles via constants like PRIMARY_DEFAULT_MODEL and preferredModels (imported by ModelCombobox.tsx, AutoFixConfigForm.tsx, OrganizationProvidersAndModelsPage.tsx, etc.). Future implementations of isFreeModel will need to consult server-only resources (e.g. a Redis-backed membership set), and folding those imports into models.ts would drag the relevant clients into client bundles.

Move the function to its own apps/web/src/lib/ai-gateway/is-free-model.ts. No behavior change; body is identical. isOpenRouterStealthModel becomes exported so the new file can use it.
