+++ apps/web/src/lib/ai-gateway/models.ts
@@ -58,16 +58,6 @@ export const preferredModels = [
   GLM_CURRENT_MODEL_ID,
 ];
 
-export async function isFreeModel(model: string): Promise<boolean> {
-  return (
-    isKiloExclusiveFreeModel(model) ||
-    model === KILO_AUTO_FREE_MODEL.id ||
-    (model ?? '').endsWith(':free') ||
-    model === 'openrouter/free' ||
-    isOpenRouterStealthModel(model ?? '')
-  );
-}
-
 export function isPdfSupportingModel(model: string): boolean {
   return isClaudeModel(model) || isOpenAiModel(model) || isGrok4Model(model);
 }
@@ -97,7 +87,7 @@ export function isKiloStealthModel(model: string): boolean {
   return kiloExclusiveModels.some(m => m.public_id === model && m.flags.includes('stealth'));
 }
 
-function isOpenRouterStealthModel(model: string): boolean {
+export function isOpenRouterStealthModel(model: string): boolean {
   return model.startsWith('openrouter/') && (model.endsWith('-alpha') || model.endsWith('-beta'));
 }
